+++ CotEditor/Sources/TextFind.swift
@@ -54,6 +54,7 @@ final class TextFind {
         
         case regularExpression(reason: String)
         case emptyFindString
+        case emptyInSelectionSearch
         
         
         var errorDescription: String? {
@@ -63,6 +64,8 @@ final class TextFind {
                 return "Invalid regular expression".localized
             case .emptyFindString:
                 return "Empty find string".localized
+            case .emptyInSelectionSearch:
+                return "The option “in selection” is enabled, although nothing is selected.".localized
             }
         }
         
@@ -74,6 +77,8 @@ final class TextFind {
                 return reason
             case .emptyFindString:
                 return "Input text to find.".localized
+            case .emptyInSelectionSearch:
+                return "Select the search scope in document or disable “in selection” option.".localized
             }
         }
         
@@ -117,6 +122,10 @@ final class TextFind {
         
         guard !findString.isEmpty else {
             throw TextFind.Error.emptyFindString
+        }
+        
+        guard !inSelection || selectedRanges.contains(where: { !$0.isEmpty }) else {
+            throw TextFind.Error.emptyInSelectionSearch
         }
         
         switch mode {
@@ -220,6 +229,8 @@ final class TextFind {
     ///   - count: The total number of matches in the scopes.
     ///   - wrapped: Whether the search was wrapped to find the result.
     func findInSelection(forward: Bool) -> (range: NSRange?, count: Int, wrapped: Bool) {
+        
+        assert(self.inSelection)
         
         var matches = [NSRange]()
         self.enumerateMatchs(in: self.selectedRanges, using: { (matchedRange, _, _) in